Misc. Camera tweaks

+++ Engine/AM2E/Camera.cs
@@ -7,22 +7,22 @@ public static class Camera
 {
     // I have NO CLUE how matrices work
     // ...but this guy does: https://www.youtube.com/watch?v=ceBCDKU_mNw
-    public static int X { get; private set; } = 0;
-    public static int Y { get; private set; } = 0;
+    public static float X { get; private set; } = 0;
+    public static float Y { get; private set; } = 0;
     public static Matrix Transform { get; private set; }
     public static int Width => Renderer.GameWidth;
     public static int Height => Renderer.GameHeight;
-    public static int BoundLeft => X - Width / 2;
-    public static int BoundRight => X + Width / 2;
-    public static int BoundTop => Y - Height / 2;
-    public static int BoundBottom => Y + Height / 2;
+    public static int BoundLeft => (int)X - Width / 2;
+    public static int BoundRight => (int)X + Width / 2;
+    public static int BoundTop => (int)Y - Height / 2;
+    public static int BoundBottom => (int)Y + Height / 2;
 
     static Camera()
     {
         UpdateTransform();
     }
 
-    public static void SetPosition(int x, int y)
+    public static void SetPosition(float x, float y)
     {
         X = x;
         Y = y;
